Refactor cloud backup write into modules
Split the cloud backup write actor into dedicated submodules and clean up related logic.

- Added client.rs and types.rs and moved CloudBackupWriteClient and related types (CloudBackupUploadedWallet, CloudBackupWriteCompletion, CloudBackupUploadedWalletsStateMode, CloudBackupWalletCountRefresh) out of supervisor.rs into types.rs and client.rs.
- Updated write.rs to re-export the new client and types and to expose supervisor types.
- Simplified await handling in rust/src/manager/cloud_backup_manager.rs (chained into_result) and consolidated await/result validation in the new client await_result (checks operation origin mismatches).
- Modified supervisor.rs to import and use the new types, removed duplicated type definitions, and adjusted write completion handling so local state is applied only after remote write succeeds and persisted enabled state is saved before marking uploaded wallets pending confirmation.
- Minor API/usage changes: updated wallet field accessors (wallet_id(), record_id()/revision_hash() to_owned()) and small control-flow adjustments.
- Tests updated to project exclusive operation lifecycle in ops/tests.rs (project_exclusive_operation_started/finished) to reflect operation-scoped writes.
- Verification flow changes in verify/session.rs: MasterKeyResolution now differentiates continuation vs finished outcomes (Continue(ContinuableMasterKeyResolution) / Finished), enabling early exit or continuation more clearly.

Overall this refactor improves module separation, correctness of operation-origin checks, and ordering of local state persistence relative to remote writes.
