+++ rust/src/manager/cloud_backup_manager/actors/write.rs
@@ -1,10 +1,16 @@
+mod client;
 mod supervisor;
+mod types;
 mod worker;
 
+pub(crate) use self::client::CloudBackupWriteClient;
 pub(crate) use self::supervisor::{
+    CloudBackupWriteBlocker, CloudBackupWriteCommandResult, CloudBackupWriteResultReceiver,
+    CloudBackupWriteSupervisor,
+};
+pub(crate) use self::types::{
     CloudBackupUploadedWallet, CloudBackupUploadedWalletsStateMode, CloudBackupWalletCountRefresh,
-    CloudBackupWriteBlocker, CloudBackupWriteClient, CloudBackupWriteCommandResult,
-    CloudBackupWriteCompletion, CloudBackupWriteResultReceiver, CloudBackupWriteSupervisor,
+    CloudBackupWriteCompletion,
 };
 
 pub(crate) type CloudBackupWriteError = crate::manager::cloud_backup_manager::CloudBackupError;
